requirements.txt
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
pytesseract==0.3.10
streamlit==1.16.0
